+++ compiler/semcall.nim
@@ -196,7 +196,7 @@ proc presentFailedCandidates(c: PContext, n: PNode, errors: CandidateErrors):
   # argument in order to remove plenty of candidates. This is
   # comparable to what C# does and C# is doing fine.
   var filterOnlyFirst = false
-  if optShowAllMismatches notin c.config.globalOptions and not isDefined(c.config, "nimPreviewTypeMismatch"):
+  if optShowAllMismatches notin c.config.globalOptions and not isDefined(c.config, "nimConciseTypeMismatch"):
     for err in errors:
       if err.firstMismatch.arg > 1:
         filterOnlyFirst = true
@@ -213,7 +213,7 @@ proc presentFailedCandidates(c: PContext, n: PNode, errors: CandidateErrors):
       inc skipped
       continue
 
-    if isDefined(c.config, "nimPreviewTypeMismatch"):
+    if isDefined(c.config, "nimConciseTypeMismatch"):
       candidates.add "[" & $err.firstMismatch.arg & "] "
 
     if err.sym.kind in routineKinds and err.sym.ast != nil:
@@ -225,7 +225,7 @@ proc presentFailedCandidates(c: PContext, n: PNode, errors: CandidateErrors):
     candidates.add("\n")
     let nArg = if err.firstMismatch.arg < n.len: n[err.firstMismatch.arg] else: nil
     let nameParam = if err.firstMismatch.formal != nil: err.firstMismatch.formal.name.s else: ""
-    if n.len > 1 and not isDefined(c.config, "nimPreviewTypeMismatch"):
+    if n.len > 1 and not isDefined(c.config, "nimConciseTypeMismatch"):
       candidates.add("  first type mismatch at position: " & $err.firstMismatch.arg)
       # candidates.add "\n  reason: " & $err.firstMismatch.kind # for debugging
       case err.firstMismatch.kind
@@ -305,7 +305,7 @@ proc notFoundError*(c: PContext, n: PNode, errors: CandidateErrors) =
   result.add(describeArgs(c, n, 1, prefer))
   result.add('>')
   if candidates != "":
-    if isDefined(c.config, "nimPreviewTypeMismatch"):
+    if isDefined(c.config, "nimConciseTypeMismatch"):
       result.add("\n" & errExpectedPosition & "\n" & candidates)
     else:
       result.add("\n" & errButExpected & "\n" & candidates)